services/llm/pyproject.toml
[project]
name = "llm"
version = "0.1.0"
description = "LLM service with LangGraph deterministic pipeline (clarify/generate/repair)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "grpcio-reflection>=1.68.0",
    "langgraph>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langchain-core>=0.3.0",
]
